upgrade the code from g++ 3.3.5 to 3.4.5: templated, derived classes must use this-> to access members of their base class abs() arguments must be cast to int NOTE: some of this was already done in r1182 and was overwritten in r1217
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@1264 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/MPProblem/Boundary.cpp
@@ -116,8 +116,8 @@ GetClearance(Vector3D point3d) const {
   double min_clearance = -1;
   double clearance = 0;
   for (int i = 0; i < pos_dofs; ++i) {
-    clearance = min(abs( point3d[i] - bounding_box[i].first), 
-			abs( bounding_box[i].second - point3d[i]));
+    clearance = min(abs((int)( point3d[i] - bounding_box[i].first)), 
+			abs((int)( bounding_box[i].second - point3d[i])));
     if (clearance < min_clearance || i == 0)
       min_clearance = clearance;
   }
